fix: resolve all clippy warnings and dead code in screenpipe-audio
- Gate RestartCooldown struct/impl with #[cfg(test)] (only used in tests)
- Fix unused variables: chunk, i, prev_idx → prefix with _
- Fix loop variables used only for indexing → use iterator/slice patterns
- Remove unnecessary u32→u32 casts for SAMPLE_RATE
- Collapse nested if statements in dedup_benchmark
- Replace clone() with std::slice::from_ref in cross_device_benchmark
- Use RangeInclusive::contains() instead of manual comparison
- Remove empty lines after doc comments in test files

+++ crates/screenpipe-audio/tests/audio_pipeline_benchmark/quality_regression.rs
@@ -45,7 +45,7 @@ async fn whisper_hallucination_on_silence() {
 
     let text = screenpipe_audio::stt(
         &silence,
-        SAMPLE_RATE as u32,
+        SAMPLE_RATE,
         "test-silence",
         engine.clone(),
         None,
@@ -91,7 +91,7 @@ async fn whisper_hallucination_on_noise() {
 
     let text = screenpipe_audio::stt(
         &noise,
-        SAMPLE_RATE as u32,
+        SAMPLE_RATE,
         "test-noise",
         engine.clone(),
         None,
@@ -135,7 +135,7 @@ async fn whisper_hallucination_on_ambient() {
 
     let text = screenpipe_audio::stt(
         &ambient,
-        SAMPLE_RATE as u32,
+        SAMPLE_RATE,
         "test-ambient",
         engine.clone(),
         None,
@@ -240,7 +240,7 @@ async fn whisper_real_speech_still_works() {
 
             let text = screenpipe_audio::stt(
                 chunk,
-                SAMPLE_RATE as u32,
+                SAMPLE_RATE,
                 "benchmark",
                 engine.clone(),
                 None,
